requirements.txt
asgiref==3.5.2
astroid==2.12.12
asttokens==2.2.1
attrs==22.1.0
backcall==0.2.0
black==22.10.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cloudinary==1.30.0
decorator==5.1.1
dill==0.3.6
dj-database-url==1.2.0
Django==4.0.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
exceptiongroup==1.0.4
executing==1.2.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.7.0
isort==5.10.1
jedi==0.18.2
jsonschema==4.17.3
lazy-object-proxy==1.8.0
matplotlib-inline==0.1.6
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==22.0
parso==0.8.3
pathspec==0.10.2
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.5.4
pluggy==1.0.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pyflakes==2.2.0
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==7.2.0
pytest-django==4.5.2
pytest-testdox==3.0.1
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
traitlets==5.7.0
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
